#1178 completed migration of lynx functions to ch23_lynx
