+++ src/ch23_lynx/test/lynx/test_etl_spark_person_csvs_to_lesson_jsons.py
@@ -6,7 +6,7 @@ from ch11_bud._ref.ch11_path import (
     create_spark_all_lesson_path as all_lesson_path,
 )
 from ch18_etl_config.etl_sqlstr import create_prime_tablename
-from ch19_etl_steps.etl_main import etl_spark_person_csvs_to_lesson_json
+from ch23_lynx.lynx_main import etl_spark_person_csvs_to_lesson_json
 from os.path import exists as os_path_exists
 from ref.keywords import Ch23Keywords as kw, ExampleStrs as exx
