+++ src/ch23_lynx/test/lynx/test_moment_sparks_lesson_jsons_to_personunit_json.py
@@ -7,7 +7,7 @@ from ch11_bud._ref.ch11_path import (
     create_spark_all_lesson_path,
     create_spark_expressed_lesson_path,
 )
-from ch19_etl_steps.etl_main import etl_spark_lesson_json_to_spark_inherited_personunits
+from ch23_lynx.lynx_main import etl_spark_lesson_json_to_spark_inherited_personunits
 from os.path import exists as os_path_exists
 from ref.keywords import Ch23Keywords as kw, ExampleStrs as exx
